pyproject.toml
[project]
name = "sourcesense"
version = "0.1.0"
dependencies = [
    "atlan-application-sdk",
    "dapr>=1.15.0",
    "fastapi",
    "uvicorn",
    "httpx",
    "orjson>=3.11.3",
    "pandas>=2.3.2",
    "temporalio>=1.17.0",
    "python-dotenv",
    "Jinja2",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "respx",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.uv]
# This section is reserved for uv-specific configurations.
